Remove dynamoDB delete table until localstack fix it

+++ Tests/AWSSDKSwiftTests/Services/Dynamodb/DynamoDBTests.swift
@@ -71,9 +71,14 @@ class DynamoDBTests: XCTestCase {
         }
     }
     
-    func deleteTable(name: String) -> EventLoopFuture<DynamoDB.DeleteTableOutput> {
+    func deleteTable(name: String) -> EventLoopFuture<Void> {
+        // for some reason delete table is not working while running within a github action. works everywhere else
+        // have verified this is an issue with awscli as well, so not an issue with aws-sdk-swift
+        guard ProcessInfo.processInfo.environment["GITHUB_ACTIONS"] != "true"  else {
+             return Self.dynamoDB.client.eventLoopGroup.next().makeSucceededFuture(())
+        }
         let input = DynamoDB.DeleteTableInput(tableName: name)
-        return Self.dynamoDB.deleteTable(input)
+        return Self.dynamoDB.deleteTable(input).map { _ in }
     }
     
     func putItem(tableName: String, values: [String: Any]) -> EventLoopFuture<DynamoDB.PutItemOutput> {
